+++ common/game/game.go
@@ -12,7 +12,7 @@ type State interface {
 	IsGoalState(bool) (bool, interface{})
 	EvaluateGoalState() float64
 	Same(State) bool
-	GenSample(float64, chan []uint64, chan [2][]int) string // Returns a string representing state attributes for supervised machine learning
+	GenSample(float64, chan []uint32, chan [2][]int) string // Returns a string representing state attributes for supervised machine learning
 }
 
 // ------------------
